Fix a quorum merge strategy invariant breakage
Replay diverges when it doesn't match canon, *always*. Until now there
was this unclear state where it doesn't when it's shorter, except it
does permanentely once it's longer.

No more of this. When asked if it diverges, always check if it diverges.
Rest of code should be fine on invariants and doesn't check for
divergence needlessly so there's no worrying about performance.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>