Use a more correct sync primitive
'Notify' behaves in odd ways we don't want. It can return immediately if
we notified before waiting, it might not return if we notify right after
checking a value, but before awaiting.
So, we use watch instead and put some more care into ensuring it works
as we want it to.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>